requirements.txt
# Core Flask application
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.0
python-socketio==5.8.0
python-engineio==4.7.1

# Database and caching
supabase==2.0.0
httpx==0.23.0
psycopg2-binary==2.9.7

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# External APIs
requests==2.31.0
spotipy==2.23.0

# Data processing
python-dateutil==2.8.2
pytz==2023.3

# Astronomy and astrology calculations
ephem==4.1.4

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
python-dotenv==1.0.0
